Added encryption, connected MQTT to database, code cleanup

+++ Database/init.sql
@@ -7,10 +7,10 @@ CREATE TABLE userdata.logins (
 CREATE SCHEMA chengeta;
 CREATE TABLE chengeta.sounds (
     time timestamp NOT NULL,
-    nodeid INT NOT NULL,
-    latitude INT NOT NULL,
-    longitude INT NOT NULL,
+    nodeid BIGINT NOT NULL,
+    latitude double precision NOT NULL,
+    longitude double precision NOT NULL,
     soundtype VARCHAR(200) NOT NULL,
-    probability INT NOT NULL,
+    probability SMALLINT NOT NULL,
     soundfile VARCHAR(500) NOT NULL                         
 );
